Fix MyPy type errors in experiment scripts
- Convert np.mean() results to float() in leniency_experiment.py and collusive_leniency_experiment.py
- Resolves dict type mismatch: numpy.floating[Any] vs float
- Ensures GitHub workflow passes all type checks

+++ scripts/collusive_leniency_experiment.py
@@ -420,8 +420,8 @@ def run_collusive_leniency_experiment(
             "avg_consumer_surplus": float(
                 np.mean([ep["consumer_surplus"] for ep in episodes])
             ),
-            "avg_producer_surplus": np.mean(
-                [ep["producer_surplus"] for ep in episodes]
+            "avg_producer_surplus": float(
+                np.mean([ep["producer_surplus"] for ep in episodes])
             ),
         }
